+++ src/test/kotlin/com/kotlin/boardproject/service/TimeTableServiceImplTest.kt
@@ -197,14 +197,3 @@ class TimeTableServiceImplTest : BehaviorSpec(
         }
     },
 )
-
-fun setScheduleTableRepository(
-    timeTableOne: TimeTable,
-    timeTableTwo: TimeTable,
-    timeTableThree: TimeTable,
-    scheduleRepository: ScheduleRepository,
-) {
-    every { scheduleRepository.findByTimeTableFetchDayOfWeekTimePairs(timeTableOne) } returns timeTableOne.schedules
-    every { scheduleRepository.findByTimeTableFetchDayOfWeekTimePairs(timeTableTwo) } returns timeTableTwo.schedules
-    every { scheduleRepository.findByTimeTableFetchDayOfWeekTimePairs(timeTableThree) } returns timeTableThree.schedules
-}
